requirements.txt
ckanapi
requests>=2.4.2
sqlalchemy
pytest
pytest-ckan
